Schema changes required for cobbler integration (by no means definitive).  In general, fields added *do* need WUI visibility at first, but later things like profile import will start to include them in the specfiles (once we have specfiles).
The act of deploying a system, for instance, will update the cobbler config based on the database and then deploy that system using koan.  We might also care about provisioning bare metal to run these virtual containers, which is why there are also fields added to machine.  (Also, bare metal can be appliances).

db2cobbler functionality is TBA as is the backend code to surface this data, for now they are left as NULL.

note that the demodata.sql file has been emptied as the SQL was wrong given the new handling of UID's.  Should we want to add that back, the SQL can be fixed at a later time.

As of this commit, or any that changes the schema, deletion of "primary_db" and re-running of "setup_db.sh" and "populate.sh" is a good idea.