In the middle of adding cobbler bootstrapping code.  Out of the box (from an RPM install) the web service will run but will not let users log in.  It will return ERR_MISCONFIGURED.  If the WUI gets this as a return, it should show a splash screen telling the user to run "shadow init", edit the config files, and then run "shadow import" server side, and then reload the page.  This is the install process and "shadow init" will probably also include other setup things in the future.
Bootstrapping works by first doing an import of a mirror (rsync only for now), and then importing the distributions
into cobbler.  From the cobbler distro list, the database list is populated, and then from then on the database
generates the cobbler config and not vice versa.  The idea is that cobbler is to be a "Requires" of the RPM, but the user
should *not* have to understand cobbler -- it's used as a library.  Shadowmanager will "take over" a cobbler installation
and manage it, not the other way around.

This will be more apparent as the integration work continues.  So far the only change needed to be made to the WUI
is to recognize ERR_MISCONFIGURED.  In addition, the integration of cobbler WILL mean that validation on the backend
will become *more* strict, and until validation is made error-tight, more ERR_UNCAUGHT_EXCEPTIONS will be thrown
as opposed to ERR_INVALID_ARGUMENTS.  The intent is to trap anything that makes cobbler get confused and
stop it prior to firing off cobbler.  This also means some data transmogrification across the ShadowManager/Cobbler
boundary.