Most of the refactoring is here.
Basic changes include moving all the export api to being class methods of
classes derived from the web_svc.WebSvc class.

The web_svc.WebSvc class includes api registration functions, shared db
handles, logging, and config hooks. All of the exposed api handlers should
be subclasses of this.

web_svc.WebSvc includes basic authentication support for all the methods
that needs auth (say, everything but registration).

shadow.py should now be mostly ignorant of the exposed api. The
wrapper methods for each exposed method are now gone. Each
api modules is imported and it's registration methods invoked
to discover the export api methods.

Each class maps the external api name with the internal api
in the self.methods dictionary. At the moment, I've kept the
existing flat hierarchy (aka, server.deployment_add, etc) but
in the future it should be possible to change it to say,
server.deployment.add instead.

Most of the api modules (aka, modules/*.py) now include
a SplatData class and a Splat class. The SplatData class
is typically the data struct and it's various marshalling
magic, while the Splat class is the exposed api.

Most of the api modules that need to invoke methods on other
api modules should now be doing that directly, instead of
going though the xmlrpc stack.

Other general changes include a slight reorg of imports.

We now how an internal "config_data" class, that the app itself
uses for it's config, and a modules/config.py that implements the
exposed api that was previously available. Modules that need to
use the config should import config_data not modules/config
The config object is also implemented as a Singleton, so
there shouldn't be any config issues.