add some improved debugging support
try to log a stack trace and some other basic info on calls that raise exceptions
all the way back to the handler.

It needs some work, but this should help trouble shoot stuff for now.

FIXME: the code currently grabs a snap of the stack trace when the
exceptions are created. The way we currently use exceptions, this
is probably suboptimal