Work to get the db helper functions in line (and associated work on tasks)
Add -- works fine, though later we'll need to add awareness of which user created the action and
this will be inferred from the token.  Null handling in the DB code is a bit odd, though I think
in the future we'll have less nullable fields to deal with.

Delete -- works fine

List ok -- for view, some fields still show as blank

Edit -- display needs tweaking.