Rough start of a "design doc", though it probably shouldn't be called that.
Here to give a basic overview of terminology and concepts, and how objects
relate, so we can (A) verify we're all on the same page when discussing
concepts and (B) have a place to document how the various pieces are going
to interoperate.

Work in progress, by no means final, please edit as needed.