sm_get_puppet_node.py retrieves a list of puppet classes for a given node, as required by puppetmaster for external node storage.
Output format follows the expectations of puppetmaster -- first line is the parent node (currently always blank for shadowmanger stuff). The next line is a space-separated list of puppet classes which apply to this node.

Currently, the classes are retrieved by pulling the list of puppet classes for the image associated with the given node -- the node is determined by comparing the given nodename with the address listed for the machine.

Not yet supported are deployments or per-node overrides of the default puppet classes.