Miscellaneous changes related to get some things broken as a function of earlier commits working.
Newest changes also include bugfixes down the path of registrating without a token (previously not heavily tested) and what happens when a machine doesn't initially have a profile assigned to it.

From my testing, import definitely works again, as does registration, but at this point editing a machine causes it's MAC addr to be set to null.  Basically it loses the registration bits.