Checking in remander of fixes, this actually causes some weirdness with machine list.  What I think is happening is that the WUI is not setting the attributes for fields that are coming up null, and they are being deferenced later.  The current errors are showing up in the view code.
In general, what the WUI needs to do is detect the retrieve_all results
with negative ID's and flush those out, somewhere in ManagedObject, so they
are not listed.

I am not sure ATM what's up with the view code and the new templating
errors, but it's looking a lot better, especially in regards to the
contents of svclog.  The new db_util stuff seems to be returning the right
results as I would expect them to be returned.

There could be some legacy "treat -1 as null" code in the WUI somewhere,
and if so, that needs to be pulled.  Same for backend.