Virt management and registration changes.
Registration now saves several files in /etc/sysconfig/virtfactory that the
node daemon can come back to read later.

Lots of virt management code coming in, still untested, but gives the
basics of what we'll be doing.  We're using the fact that koan saves
it's config files and image files in a known location to bootstrap a lot
of the libvirt management from those files.

TBD: a call to a node to update ip and other info would be useful, though
most of our contacts are using hostnames anyway, as is puppet.