Two things.
(A) update the WUI task/edit page so it can be used programmatically (even
if not linked) to feed taskatron from the WUI.

(B) add a configurable setting in the settings file for adjusting log
level.  Defaults to "INFO".  Turning on DEBUG with taskatron is quite
undesirable as it will result it 300MB log files.  Which probably means
having a seperate config file for taskatron will soon be useful.