This is ongoing work to move the startup scripts to Postgres.
It turns out that the "upgrade init" scripts we have earlier need root,
which doesn't currently work in the DB as we have to be sudo'd to postgres
to get the scripts to run.

We can probably fix this by either (A) tweaking postgres or (B) making
the upgrade init scripts just use SQL and not have dependencies on the
server modules which need root.