using puppet modules properly for vf_import now. vf_get_puppet_node now uses the xmlrpc interface rather than calling the modules directly (since that no longer works if you aren't root). We'll need to work through the security implications of exposing the node mapping without auth. We may want to plug the puppet node script into the auth framework somehow or otherwise limit the puppet node api to localhost only.
The test1, etc. profiles don't currently work since the puppet class
mapping doesn't match the defined manifest, although until we have
something more than "hello world" there it probably doesn't matter. I
updated container to at least match the helloworld? classes for testing
purposes.

The other remaining problem is that currently to get the fileserver mapping
to work I have to set $server manually in site.pp. We can automate this as
part of vf_import or in virt-factory-server install, but hopefully there's
a way of doing this directly in puppet without having to set the server
hostname in the manifest.