profiles are now installed as RPMs. to generate the RPM, type 'make' in the profiles/foo directory as before. when the RPM is installed, vf_import is called in %post -- the RPM puts the puppet manifest/files in place, so vf_import no longer extracts a tarball -- it just does the site-wide manifes munging and the db imports.
also new is vf_gen_profile_stub -- automates most of the profile generation
process. It generates the makefile, specfile, and profile.xml and creates a
blank init.pp. All of the attributes in profile.xml can be entered on the
cmdline if desired, or the file can be edited directly later. The only
required option is the profile name.

Another functional change is that we're no longer downcasing the profile
name when generating the puppet module (and thus the directory names) --
the reason for this is that puppet module names are case sensitive, so we
should preserver the case of the specified profile name.