Changes to make virt factory DB creation work with sqlalchemy, plus changes in the init scripts.
Database name is 'virtfactory', which is dropped and recreated when using
--replace.  --replace needs the postgres user.

PG authentication is "repaired" when run with "--fixauth".  This command
needs root access.

No optparsing ATM.

Also, we no longer care where the db is, so notice the check for
/etc/virtfactory/db/exists, which is created after a successful creation.
This is how the init scripts knows whether to upgrade or not.  This can
probably be made cleaner.

See init scripts for examples on when what gets invoked.